+++ src/main/resources/db/migration/V1000__laptops_tablets_phones.sql
@@ -1,29 +1,29 @@
 create table laptops (
-     id serial not null,
+     id bigserial not null,
      name varchar(128) not null,
      image varchar(128) not null,
      quantity integer not null,
      description varchar(128) not null,
-     price INTEGER not null,
+     price FLOAT not null,
      primary key (id)
 );
 
 create table tablets (
-     id serial not null,
+     id bigserial not null,
      name varchar(128) not null,
      image varchar(128) not null,
      quantity integer not null,
      description varchar(128) not null,
-     price INTEGER not null,
+     price FLOAT not null,
      primary key (id)
 );
 
 create table phones (
-    id serial not null,
+    id bigserial not null,
     name varchar(128) not null,
     image varchar(128) not null,
     quantity integer not null,
     description varchar(128) not null,
-    price INTEGER not null,
+    price FLOAT not null,
     primary key (id)
 );
